Rename: TokenManager 추상체 및 구현체 이름변경
