+++ core-database/src/main/java/com/hmoa/core_database/TokenManager.kt
@@ -1,71 +1,16 @@
 package com.hmoa.core_database
 
-import androidx.datastore.core.DataStore
-import androidx.datastore.preferences.core.Preferences
-import androidx.datastore.preferences.core.edit
-import androidx.datastore.preferences.core.stringPreferencesKey
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.map
-import javax.inject.Inject
 
-class TokenManager @Inject constructor(private val dataStore: DataStore<Preferences>) : TokenManagerInterface {
-    companion object {
-        private val AUTH_TOKEN_KEY = stringPreferencesKey("AUTH_TOKEN")
-        private val REMEMBERED_TOKEN_KEY = stringPreferencesKey("REMEBERED_TOKEN")
-        private val KAKAO_ACCESS_TOKEN_KEY = stringPreferencesKey("KAKAO_ACCESS_TOKEN")
-    }
+interface TokenManager {
 
-    override fun getAuthToken(): Flow<String?> {
-        return dataStore.data.map { preferences ->
-            preferences[AUTH_TOKEN_KEY]
-        }
-    }
-
-    override fun getRememberedToken(): Flow<String?> {
-        return dataStore.data.map { preferences ->
-            preferences[REMEMBERED_TOKEN_KEY]
-        }
-    }
-
-    override fun getKakaoAccessToken(): Flow<String?> {
-        return dataStore.data.map { preferences ->
-            preferences[KAKAO_ACCESS_TOKEN_KEY]
-        }
-    }
-
-    override suspend fun saveAccessToken(token: String) {
-        dataStore.edit { preferences ->
-            preferences[AUTH_TOKEN_KEY] = token
-        }
-    }
-
-    override suspend fun saveRememberedToken(token: String) {
-        dataStore.edit { preferences ->
-            preferences[REMEMBERED_TOKEN_KEY] = token
-        }
-    }
-
-    override suspend fun saveKakaoAccessToken(token: String) {
-        dataStore.edit { preferences ->
-            preferences[KAKAO_ACCESS_TOKEN_KEY] = token
-        }
-    }
-
-    override suspend fun deleteAccessToken() {
-        dataStore.edit { preferences ->
-            preferences.remove(AUTH_TOKEN_KEY)
-        }
-    }
-
-    override suspend fun deleteRememberedToken() {
-        dataStore.edit { preferences ->
-            preferences.remove(REMEMBERED_TOKEN_KEY)
-        }
-    }
-
-    override suspend fun deleteKakaoAccessToken() {
-        dataStore.edit { preferences ->
-            preferences.remove(KAKAO_ACCESS_TOKEN_KEY)
-        }
-    }
+    fun getAuthToken(): Flow<String?>
+    fun getRememberedToken(): Flow<String?>
+    fun getKakaoAccessToken(): Flow<String?>
+    suspend fun saveAccessToken(token: String)
+    suspend fun saveRememberedToken(token: String)
+    suspend fun saveKakaoAccessToken(token: String)
+    suspend fun deleteAccessToken()
+    suspend fun deleteRememberedToken()
+    suspend fun deleteKakaoAccessToken()
 }
